dfa: Identify states by the NFA states at entry.
Only use the set of entry states rather than the closure of states passed
through. This should be equivalent because each entry state contributes a
unique set of states to the closure.

The computation in the outgoing transition assumes the threads begin at
the entry state, so the states passed through can't be considered part of
the initial state. This also allows us to identify states before computing
the closure.