Allow a downward send in the followlast position of upwards receive
This occurs in cases like

```
repeat {
    on read(b) {
        data(CMD_READ)
        data(b)
    }
}

data(CMD_STOP)
```

where the desired semantic is to wait for an upwards token, and if it is
not `read` to exit the loop and run `data(CMD_STOP)`, leaving the token
pending to be handled by a subsequent `on` or transaction end. The
bytecode `Test` already falls through like this, so it's just removing
checks against it.